Fix hover in LazyColumn
LazyColumn doesn't use the first measureAndLayout on scroll, it uses the second one. In the first one we already make update for hover, calling `pointerPositionUpdater.needUpdate()`. We need to do so in the second one as well (and update contentSize, if it were the root elements that were changed)

Fixes https://github.com/JetBrains/compose-jb/issues/2169

Also fix the test `rendering of LazyColumn`, it fails because of some changes in AOSP. When I disable hover updating, it still fails.