Remove neediness of Android NDK, fix Windows build
Android NDK is only needed for Android, and only for Layout inspector and benchmarking. Usually, when we develop Compose Multiplatform, we rare check something on Android, and almost never in Layout inspector or in benchmarks.

Keeping Android NDK fresh on every rebase is a difficult task, because it has a non-default tree structure.

I checked, the Android target keeps working.