Show new window/dialog on the same display.
Tested manually on Windows/macOs/Linux with two displays. Tested the case when we disconnect one display.

The new behaviour of opening windows is different from platform default.
Windows/macOS, for example, shows every new window slighly shifted in compare to the previous opened window.
We on the other hand shift in compare to the previous focused window, not the opened.

RelNote:
Changed behaviour of the initial position of the new window. Now it shows on the display, where was the latest focused window

Fixes https://github.com/JetBrains/compose-jb/issues/1845