Change replacing Compose maven group and version code
This feature will be removed in https://android-review.googlesource.com/c/platform/frameworks/support/+/2313518

Because now we have our own fork, and we can configure our build as we want, there is no point to keep Jetbrains-specific build logic in AOSP, if it is only a extension of AOSP build logic, or small change of it.

But in the future, if it is huge rewrite of the build logic, or there are some obstacles, we can discuss with AOSP build team what can be improved in AOSP code.