Fix scrolling when we change LazyListState
Fixes https://github.com/JetBrains/compose-multiplatform/issues/3551

When the new state is passed to LazyList, it goes down to `Modifier.pointerScrollable` as `controller: ScrollableState`, and wraps to `rememberUpdatedState`, which is always the same instance. Because it is the same instance, MouseWheelScrollableElement.update is never called.

Passing it is as a value now, as MouseWheelScrollableElement doesn't need a state.