Fix TapGestureDetector mutex is not locked error on Touchscreen
Fixes https://github.com/JetBrains/compose-multiplatform/issues/3655

The issue is introduced after migration to `awaitPointerEventScope`, this code is indeed has a race condition, as it was assumed in the TODO. When we send Press/Release without any moves - we go via this path:
```
// awaitPointerEventScope isn't called until it receives the first event
awaitPointerEventScope {
  // in the beginning of the method we have 2 events already in the queue

  launch { pressScope.reset() } // launch async, reset isn't called
  val down = awaitPress // don't suspend here, because we already have Press
  awaitReleaseOrCancelled() // don't suspend here, because we already have Release
  pressScope.release() // crash, because `pressScope.reset()` isn't yet called
}
```

To avoid races, we should always access `pressScope` in `launch` (this way calls maintain order).

The same approach is used in the original `TapGestureDetector.kt` (the change in commit 32de9dd2)

Also, the rest of the code doesn't have races with `pressScope` calls.

## Testing
- haven't reproduced with mouse
- reproduced with a simple test (added to OnClickTest)