Remove unused code in settings.gradle
- remove non-COMPOSE modules
- remove TODO and commented code
- remove Gradle Enterprise plugin. It seems only work in internal Google environment. We may consider similar optimization in the future

In this PR I don't try to avoid merge-conflicts with new Jetpack Compose's. It is better to maintain the list of modules ourselves. If new modules are added, there will be a merge conflict, and we will see what is new, and can do conscience porting/testing of the new modules.