Allow loading font families by name
This supports both JVM-embedded fonts and system fonts, but
is only possible on the JetBrains Runtime. Other JVMs do not
have the necessary APIs required to enumerate embedded fonts,
and even worse, they don't use the right font family name on
Windows and Linux (on macOS the implementation uses system
APIs who seem to work just fine, on other OSes they manually
parse fonts, and do it wrong).

For consistency, then, we only allow loading font families by
name when running on the JetBrains Runtime.

Besides that, we also need to access the closed sun.font API
package, so the --add-opens java.desktop/sun.font=ALL-UNNAMED
JVM parameter is required for this code to work. The failure
mode is to fall back to the default font family/typeface.