Invalidate parent layer every time child layer is invalidated
Fixes https://github.com/JetBrains/compose-multiplatform/issues/4681

The issue was because of this situation:
- the parent RenderNodeLayer is drawn and cached it's `picture`
- the child RenderNodeLayer isn't drawn, but `invalidate` is called

the child's `invalidate` should invalidate parent layer unconditionally, but we called it only if the child was drawn (picture != null), which is incorrect.

It is a regression after https://github.com/JetBrains/compose-multiplatform-core/pull/1260, but it just revealed the bug, not introduced it.

## Testing
- new added tests fail before, work after the fix
- the issue is no longer reproducible

This should be tested by QA

## Release Notes
> Fixes - Multiple Platforms
- _(prerelease fix)_ Fix frozen composition with pager and text field