`copyAnnotation.sh`. `to-integration` branch is created from a wrong commit
Fixes https://youtrack.jetbrains.com/issue/CMP-7226/copyAnnotation.sh.-to-integration-branch-is-created-from-a-wrong-commit

We took the base commit between the current and `origin/integration`. But it can be old. Instead, we should take just the current commit, because we copied it, not the base one.

## Testing
Rerun `copyAnnotation.sh` from eecb9af750be5ffb945e9c92d0fab8918b1bfdc7 (The green commit in the issue)