KeyCommand.DELETE_PREV_CHAR correctly deletes chars
Before, the command was deleting to the previous character break. Now, it will correctly delete a code point instead so that diacritic marks are removed off a glyph instead of deleting the entire glyph. An exception is carved out for emojis, where the entire emoji will be deleted instead of a single code point, which would decompose emojis to their parts if deleting a single code point.

Fixes: b/395659876
Test: textField_backspace_withDiacritic
Test: textField_backspace_withEmoji
Change-Id: I9600d97eb2641bd218c6d5143753b50aa8f1624a