Fix POM when there is dependency on an artifact with redirection
Fixes https://youtrack.jetbrains.com/issue/CMP-7761/1.8.0dev2186-doesnt-work-with-JPS-and-Maven

The issue was that `pom` points to a root artifact, not to `-desktop` artifact. It happens because we applied redirection, and KGP doesn't handle it.

We already had the fix for `project` dependencies in https://github.com/JetBrains/compose-multiplatform-core/pull/1171 and https://github.com/JetBrains/compose-multiplatform-core/pull/1236, but the issue also exist if we depend on an already published library.

In this PR, we apply the fix also to external dependencies. But only by using heuristic. If heuristic fails to notice redirection, or notices a wrong redirection, integration tests should fail.

## Testing
### Method 1
1. Perform ``./gradlew publishComposeJbToMavenLocal`
2. Check that `~/.m2/repository/org/jetbrains/compose/ui/ui-desktop/9999.0.0-SNAPSHOT/ui-desktop-9999.0.0-SNAPSHOT.pom` contains:
```
    <dependency>
      <groupId>androidx.collection</groupId>
      <artifactId>collection-jvm</artifactId>
      <version>1.5.0-beta01</version>
      <scope>runtime</scope>
    </dependency>
    ...
    <dependency>
      <groupId>org.jetbrains.androidx.lifecycle</groupId>
      <artifactId>lifecycle-runtime-desktop</artifactId>
      <version>2.8.4</version>
      <scope>runtime</scope>
    </dependency>
```
3. `mvn install exec:java -Dexec.mainClass="MainKt" -Dkotlin.version=2.1.0 -Dcompose.version=9999.0.0-SNAPSHOT` in https://github.com/JetBrains/compose-multiplatform/tree/master/ci/templates/maven-test-project works

### Method 2
1. Perform `gradlew publishComposeJbToMavenLocal -Pcompose.platforms=jvm,android,web` for the state before fix and after the fix
2. See the diff for ~/.m2/repository/org/jetbrains`:

## Release Notes
N/A