[Cherry-pick] Fix accessibility for popups of modal view controller (#2016)
The Compose Layers view is a full screen overlay of the current content.
To do this, the Layers container view should be placed as close as
possible to the current UIWindow.
If another view controller is presented modaly, UIWindow changes
accessibility elements, so the Compose Layers View can't be accessed by
the iOS accessibility engine.
Find the next view in the hierarchy. This view is unique to the root and
each modal view controller.

Fixes

https://youtrack.jetbrains.com/issue/CMP-7907/iOS-Full-Keyboard-access.-Doesnt-work-correctly-inside-dialogs

## Release Notes
N/A